feat: api access scope and rework service account auth (#254)
This adds the new api access scope for ZITADEL v2.
Also, some parts of the service account authentication were changed 
and more tests added.

BREAKING CHANGE: The service account token provider
must be created with AuthOptions. The api url must not
be passed twice. However, `ApiAccess = true` must be set
if the grpc client shall be used.

BREAKING CHANGE: AuthOptions does not contain `Endpoint`
or `Audience` anymore. This was streamlined to be consistent
and the `ServiceAccount` now requires the `audience` beside
the `AuthOptions`.