Fixes the IsInRole() extension method
Before this commit, the IsInRole() extension method would return `true` as long as at least one role was provided. It returns `true`, no matter whether the user in question is a member of the respective role, or if the user isn't a member of that role.

The cause is the use of Select() in combination with Any(). Select() maps a value of an Iterable from one form to another by invoking an Action. Any() (invoked without any further arguments) returns `true` whenever there is one or more element in an Iterable.

Select() was invoked for every role to map a given role to a boolean value that indicates whether the user is a member of this role or not. It does this by invoking `principal.IsInRole(role)`. The resulting collection is a list of booleans that may be either `true` or `false`. The list may also contain only `false` values.
The final Any() only checks if there are elements in the list. Therefore, it will return `true` in all cases, except if no role was provided to begin with.

Taken apart, this is what's happening, explained as code:

```cs
// given a valid principal that has no roles assigned
// given an array roles with at least one role
public static bool IsInRole(this ClaimsPrincipal principal, params string[] roles) {
var mappedValues = roles.Select(principal.IsInRole); // mappedValues = [ false ]
        bool isNotEmpty = mappedValues.Any(); // isNotEmpty = true
        return isNotEmpty; // return true ⚠️
}
```

Instead of Select(), the method Where() should have been used. Furthermore, Any() itself does except an Action. `someEnumerable.Where(somePredicate).Any()` is the same as calling `someEnumarble.Any(somePredicate)`. This commit changes the code accordingly and provides a unit test for the method in question.